backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
resend==2.5.1
qrcode[pil]==8.0
reportlab==4.2.0
apscheduler==3.10.4
httpx==0.27.0
slowapi==0.1.9
pytest==8.3.0
pytest-asyncio==0.24.0
aiosqlite==0.20.0
